Defer execution of operations on `Rocket` until their effects will be observed.
This is a prerequisite for async on_attach fairings. `Rocket` is now a
builder wrapper around the private type `RocketInner`, with operations
being run the next time it is necessary: `launch()`, `Client::new()`, or
`inspect()`. `inspect()` returns an `Inspector<'_>`, which is analogous
to and has the methods that could be called on an `&Rocket`.